Set destruction policy for all medical records in study (#2675)
* WIP

* Finish test

* Set policy by realm

+++ pepper-apis/dsm-core/src/main/java/org/broadinstitute/dsm/model/TissueList.java
@@ -79,7 +79,7 @@ public class TissueList {
     public static List<TissueList> getAllTissueListsForRealm(String realm, String query) {
         List<TissueList> results = new ArrayList<>();
         HashMap<Long, Tissue> tissues = new HashMap<>();
-        HashMap<Long, OncHistoryDetail> oncHistoryDetailHashMap = new HashMap<>();
+        HashMap<Integer, OncHistoryDetail> oncHistoryDetailHashMap = new HashMap<>();
         HashMap<String, Participant> participantHashMap = new HashMap<>();
         SimpleResult result = inTransaction((conn) -> {
             SimpleResult dbVals = new SimpleResult();
